Add ignore and comment, refactor realloc
Add dalloc_ignore() to ignore a pointer in memory leak check, it can be
useful for a daemon because the program never stop.

Add dalloc_comment() to add small comment to pointers so it's easier to
figure out what they are from dalloc error message.

Remove function attributes but keep destructor for dalloc_check_all().

Internal:

Add find_pointer_index() to search for a pointer in pointers, it search
from the last to the first pointer added.

Change __realloc() to use realloc() instead of __free() and __malloc(),
this is needed to keep comment and ignored field easily, also it should
be more efficient.